requirements.txt
numpy>=1.14.0
torch>=1.6
gdown>=3.10.1
torchvision
pillow>=8.0
onnx
onnxruntime
onnxconverter-common
# For GPU, use: onnxruntime-gpu